job tábla + entitás +dto + kezdeti conroller

+++ BACKEND/Szupercsapat-RFT/src/main/java/com/rft/repos/SchoolRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.rft.entities.School;
 
-public interface SchoolRepository extends JpaRepository<School, Long> {
+public interface SchoolRepository extends JpaRepository<School, Integer> {
 	
 }
