+++ BACKEND/Szupercsapat-RFT/src/main/java/com/rft/repos/WorkPlaceRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.rft.entities.School;
 import com.rft.entities.WorkPlace;
 
-public interface WorkPlaceRepository extends JpaRepository<WorkPlace, Long> {
+public interface WorkPlaceRepository extends JpaRepository<WorkPlace, Integer> {
 	
 }
